+++ .gitignore
@@ -19,6 +19,7 @@ __pycache__/*
 .pydevproject
 .settings
 .idea
+*.iml
 tags
 
 # Package files
