+++ tests/unit/commands/test_about.py
@@ -7,16 +7,6 @@ from aleph_client.__main__ import app
 runner = CliRunner()
 
 
-def test_about_help():
-    result = runner.invoke(
-        app, ["about", "--help"]
-    )
-
-    assert result.exit_code == 0, result.stdout
-
-    assert "version" in result.stdout
-
-
 def test_about_version():
     result = runner.invoke(
         app, ["about", "version"]
